coop/templates/project_template/requirements.txt
# our apps
django-coop
coop_cms
coop-geo
coop-tag
coop-bar
djaloha
coop-colorbox

# git+git://github.com/quinode/coop-colorbox.git
# git+git://github.com/quinode/djaloha.git
# git+git://github.com/quinode/coop-bar.git
# git+git://github.com/quinode/coop-tag.git
# git+git://github.com/quinode/coop-geo.git
# git+git://github.com/quinode/coop_cms.git
# git+git://github.com/quinode/django-coop.git

git+git://github.com/Nimn/geo-django-fla.git

django-autoslug
git+git://github.com/mozilla/unicode-slugify.git

# coop-tag deps
# django-taggit
# hg+https://bitbucket.org/quinode/django-taggit-autosuggest
# git+https://github.com/quinode/django-taggit-templatetags

# forks we depend on
hg+https://bitbucket.org/psyton/django-admin-tools
git+git://github.com/jonasgeiregat/django-templatetag-sugar.git

# our forks
hg+https://bitbucket.org/quinode/django-admintools-bootstrap/@0.0.2-alpha
git+git://github.com/quinode/uriredirect.git

git+git://github.com/quinode/django-extensions.git@use_admin_jquery

# do we still need the fork ?
git+git://github.com/quinode/django-chosen.git

# setup.py was broken
git+git://github.com/quinode/djangoembed.git

# coop deps
psycopg2
Pillow
south
rq
django-forms-builder
git+git://github.com/ui/django-rq.git
python-dateutil
sorl-thumbnail==11.09
shortuuid
django-tinymce
git+git://github.com/twidi/django-extended-choices.git
hg+https://bitbucket.org/ubernostrum/django-profiles
django-positions
django-push
mimeparse
geohash
rdflib
git+git://github.com/RDFLib/rdfextras.git
git+git://github.com/RDFLib/rdflib-jsonld.git
git+git://github.com/RDFLib/rdflib-sparqlstore.git
git+git://github.com/quinode/subhub.git
django-appconf

# webid deps
pyOpenSSL
git+git://github.com/bennomadic/django-webid-auth.git
git+git://github.com/bennomadic/django-webid-provider.git
git+git://github.com/bennomadic/django-conneg.git

# coop_cms deps
django-floppyforms==0.4.7
django-pagination
feedparser
git+git://github.com/gabrielgrant/django-html-field.git



# central logging
raven

